+++ src/services/discordAuth.service.ts
@@ -11,14 +11,13 @@ import {
 } from '../types/discordAuth.js';
 import { BusinessError, SystemError } from '../types/error.js';
 import { discordMemberRoleService } from './discordMemberRole.service.js';
+import { fetchWithTimeout } from '../utils/fetchWithTimeout.js';
 
 const discordApiBaseUrl = 'https://discord.com/api';
 const scopes = ['identify', 'guilds'];
 const clientId = process.env.DISCORD_CLIENT_ID;
 const clientSecret = process.env.DISCORD_CLIENT_SECRET;
 const redirectUri = process.env.DISCORD_REDIRECT_URI;
-
-const DISCORD_API_TIMEOUT = 10000;
 
 /**
  * @desc 최초 로그인 시 토큰 포맷
@@ -54,33 +53,6 @@ function formatRefreshedToken(tokenData: DiscordTokenAPI, oldRefreshToken: strin
  * @desc discord API 호출 및 DB 작업 처리
  */
 export class DiscordAuthService {
-  /**
-   * @desc 타임아웃이 적용된 Fetch 헬퍼 메서드
-   */
-  private async fetchWithTimeout(
-    url: string,
-    options: RequestInit,
-    timeout: number = DISCORD_API_TIMEOUT,
-  ): Promise<Response> {
-    const controller = new AbortController();
-    const timeoutId = setTimeout(() => controller.abort(), timeout);
-
-    try {
-      const response = await fetch(url, {
-        ...options,
-        signal: controller.signal,
-      });
-      return response;
-    } catch (error: any) {
-      if (error.name === 'AbortError') {
-        throw new SystemError(`Discord API Request Timed out after ${timeout}ms`, 504);
-      }
-      throw error;
-    } finally {
-      clearTimeout(timeoutId);
-    }
-  }
-
   // --- 1. Public Methods (컨트롤러에서 호출) ---
 
   /**
@@ -112,7 +84,7 @@ export class DiscordAuthService {
   ): Promise<string> {
     try {
       // 1. Discord API로 토큰 요청
-      const tokenResult = await this.fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token`, {
+      const tokenResult = await fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
         body: querystring.stringify({
@@ -131,7 +103,7 @@ export class DiscordAuthService {
       const { access_token, token_type } = tokenData;
 
       // 2. Discord API로 유저 정보 요청
-      const userResult = await this.fetchWithTimeout(`${discordApiBaseUrl}/users/@me`, {
+      const userResult = await fetchWithTimeout(`${discordApiBaseUrl}/users/@me`, {
         headers: { Authorization: `${token_type} ${access_token}` },
       });
 
@@ -208,7 +180,7 @@ export class DiscordAuthService {
    */
   public async fetchUser(accessToken: string) {
     try {
-      const userResult = await this.fetchWithTimeout(`${discordApiBaseUrl}/users/@me`, {
+      const userResult = await fetchWithTimeout(`${discordApiBaseUrl}/users/@me`, {
         headers: {
           Authorization: `Bearer ${accessToken}`,
         },
@@ -289,7 +261,7 @@ export class DiscordAuthService {
     currentToken: DiscordToken,
   ): Promise<string> {
     try {
-      const result = await this.fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token`, {
+      const result = await fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
         body: querystring.stringify({
@@ -321,7 +293,7 @@ export class DiscordAuthService {
    */
   private async revokeDiscordToken(accessToken: string): Promise<void> {
     try {
-      const result = await this.fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token/revoke`, {
+      const result = await fetchWithTimeout(`${discordApiBaseUrl}/oauth2/token/revoke`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
         body: querystring.stringify({
